requirements.txt
torch>=1.8.0
torchvision>=0.9.0
numpy>=1.19.5
Pillow>=8.1.0
tqdm>=4.62.0